\name{plot.mds}
\alias{plot.mds}
\title{
  Plot for modeled tooth dentin serial sections
}
\description{
  \code{plot.mds} is a plotting utile for quick visualization of the modeled the tooth dentin microstructures under an object of class \code{"mds"}. The modeled dentin growth layers and outer shape are plotted on a geometric plane.
}
\usage{
  \method{plot}{mds}(x, \dots)
}
\arguments{
  \item{x}{An object of class \code{"mds"} generated by \code{\link{mds}}}
  \item{\dots}{Additional arguments passed to \code{\link{plot}}.}
}
\details{
  The modeled tooth dentin is plotted on a geometric plane. Dentin horn and root tip equal the geometric origin and the coordinate (0, 1), respectively. Growth lines are shown with oblique colored lines inside the longitudinal half of tooth dentin outer shape. Horizontal sections are shown with horizontal solid black lines.
}
\references{
  Tsutaya T. in prep. Blurred time resolution of tooth dentin serial sections.
}
\author{
  Takumi Tsutaya developed this model.
}
\seealso{
  \code{\link{MDSS}}, \code{\link{mds}}
}
\examples{
## See ?mds for examples.
}
\keyword{math}
